+++ crates/full-node/sov-sequencer/tests/integration/preferred_with_reorgs.rs
@@ -234,7 +234,10 @@ async fn test_stream_of_transactions(
     .set_da_config(|da_config| {
         // We don't need to test restarts, so let's save disk accesses and file descriptors.
         da_config.connection_string = MockDaConfig::sqlite_in_memory();
-        da_config.sender_address = genesis_config.sequencer_registry.seq_da_address;
+        da_config.sender_address = genesis_config
+            .sequencer_registry
+            .sequencer_config
+            .seq_da_address;
         da_config.randomization = randomization_config;
     });