+++ crates/full-node/sov-sequencer/tests/integration/upgradability/mod.rs
@@ -390,7 +390,11 @@ async fn create_test_rollup(
     (
         new_test_rollup::<TestRuntime<TestSpec>>(
             dir.clone(),
-            genesis_params.runtime.sequencer_registry.seq_da_address,
+            genesis_params
+                .runtime
+                .sequencer_registry
+                .sequencer_config
+                .seq_da_address,
             genesis_params,
             minimum_profit_per_tx,
             true,