+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# 2025-07-17
+# 2025-07-18
+- #3246 **BREAKING CHANGE** A minimum bond is now required for sequencer registration.
+
+# 2025-07-17 
 - #3251 **BREAKING CHANGE** Update the `Module` trait so that `Module::call and Module::genesis` return `anyhow::Result<()>` instead of `Result<(), ModuleError>`.
 - #3249 This PR renames the SchemaGenerator trait to UniversalWallet, so that the trait and the macro have the same name for better DevX.